Add a facade to a tsg table

Add a facade to a tsg table

Usage

add_facade(
  data,
  table.offsetRow = 0,
  table.offsetCol = 0,
  table.gridLines = NULL,
  table.tabColour = NULL,
  table.fontName = NULL,
  table.fontSize = NULL,
  table.fontColour = NULL,
  table.bgFill = NULL,
  table.fgFill = NULL,
  table.halign = NULL,
  table.valign = NULL,
  table.wrapText = FALSE,
  table.indent = NULL,
  table.locked = NULL,
  table.hidden = NULL,
  table.decimalPrecision = NULL,
  table.decimalCols = NULL,
  table.lastRowBold = NULL,
  table.widths = NULL,
  table.widthOffset = NULL,
  title.fontName = NULL,
  title.fontSize = NULL,
  title.fontColour = NULL,
  title.border = NULL,
  title.borderColour = NULL,
  title.borderStyle = NULL,
  title.bgFill = NULL,
  title.fgFill = NULL,
  title.halign = NULL,
  title.valign = NULL,
  title.textDecoration = NULL,
  title.wrapText = NULL,
  title.textRotation = NULL,
  title.indent = NULL,
  title.heights = NULL,
  subtitle.fontName = NULL,
  subtitle.fontSize = NULL,
  subtitle.fontColour = NULL,
  subtitle.border = NULL,
  subtitle.borderColour = NULL,
  subtitle.borderStyle = NULL,
  subtitle.bgFill = NULL,
  subtitle.fgFill = NULL,
  subtitle.halign = NULL,
  subtitle.valign = NULL,
  subtitle.textDecoration = NULL,
  subtitle.wrapText = NULL,
  subtitle.textRotation = NULL,
  subtitle.indent = NULL,
  subtitle.heights = NULL,
  header.fontName = NULL,
  header.fontSize = NULL,
  header.fontColour = NULL,
  header.border = NULL,
  header.borderColour = NULL,
  header.borderStyle = NULL,
  header.bgFill = NULL,
  header.fgFill = NULL,
  header.halign = NULL,
  header.valign = NULL,
  header.textDecoration = NULL,
  header.wrapText = NULL,
  header.textRotation = NULL,
  header.indent = NULL,
  header.heights = NULL,
  spanner.fontName = NULL,
  spanner.fontSize = NULL,
  spanner.fontColour = NULL,
  spanner.border = NULL,
  spanner.borderColour = NULL,
  spanner.borderStyle = NULL,
  spanner.bgFill = NULL,
  spanner.fgFill = NULL,
  spanner.halign = NULL,
  spanner.valign = NULL,
  spanner.textDecoration = NULL,
  spanner.wrapText = NULL,
  spanner.textRotation = NULL,
  spanner.indent = NULL,
  spanner.heights = NULL,
  body.fontName = NULL,
  body.fontSize = NULL,
  body.fontColour = NULL,
  body.numFmt = NULL,
  body.border = NULL,
  body.borderColour = NULL,
  body.borderStyle = NULL,
  body.bgFill = NULL,
  body.fgFill = NULL,
  body.halign = NULL,
  body.valign = NULL,
  body.textDecoration = NULL,
  body.wrapText = NULL,
  body.textRotation = NULL,
  body.indent = NULL,
  body.heights = NULL,
  col_first.fontName = NULL,
  col_first.fontSize = NULL,
  col_first.fontColour = NULL,
  col_first.numFmt = NULL,
  col_first.border = NULL,
  col_first.borderColour = NULL,
  col_first.borderStyle = NULL,
  col_first.bgFill = NULL,
  col_first.fgFill = NULL,
  col_first.halign = NULL,
  col_first.valign = NULL,
  col_first.textDecoration = NULL,
  col_first.wrapText = NULL,
  col_first.textRotation = NULL,
  col_first.indent = NULL,
  col_first.widths = NULL,
  col_last.fontName = NULL,
  col_last.fontSize = NULL,
  col_last.fontColour = NULL,
  col_last.numFmt = NULL,
  col_last.border = NULL,
  col_last.borderColour = NULL,
  col_last.borderStyle = NULL,
  col_last.bgFill = NULL,
  col_last.fgFill = NULL,
  col_last.halign = NULL,
  col_last.valign = NULL,
  col_last.textDecoration = NULL,
  col_last.wrapText = NULL,
  col_last.textRotation = NULL,
  col_last.indent = NULL,
  col_last.widths = NULL,
  row_group.fontName = NULL,
  row_group.fontSize = NULL,
  row_group.fontColour = NULL,
  row_group.border = NULL,
  row_group.borderColour = NULL,
  row_group.borderStyle = NULL,
  row_group.bgFill = NULL,
  row_group.fgFill = NULL,
  row_group.halign = NULL,
  row_group.valign = NULL,
  row_group.textDecoration = NULL,
  row_group.wrapText = NULL,
  row_group.textRotation = NULL,
  row_group.indent = NULL,
  row_group.widths = NULL,
  source_note.fontName = NULL,
  source_note.fontSize = NULL,
  source_note.fontColour = NULL,
  source_note.border = NULL,
  source_note.borderColour = NULL,
  source_note.borderStyle = NULL,
  source_note.bgFill = NULL,
  source_note.fgFill = NULL,
  source_note.halign = NULL,
  source_note.valign = NULL,
  source_note.textDecoration = NULL,
  source_note.wrapText = NULL,
  source_note.textRotation = NULL,
  source_note.indent = NULL,
  source_note.heights = NULL,
  footnotes.fontName = NULL,
  footnotes.fontSize = NULL,
  footnotes.fontColour = NULL,
  footnotes.border = NULL,
  footnotes.borderColour = NULL,
  footnotes.borderStyle = NULL,
  footnotes.bgFill = NULL,
  footnotes.fgFill = NULL,
  footnotes.halign = NULL,
  footnotes.valign = NULL,
  footnotes.textDecoration = NULL,
  footnotes.wrapText = NULL,
  footnotes.textRotation = NULL,
  footnotes.indent = NULL,
  footnotes.heights = NULL
)

Arguments

data

A tsg table object to which the facade will be added. This is typically a data frame or tibble that has been processed using tsg functions.

table.offsetRow

Row offset of the table

table.offsetCol

Column offset of the table

table.gridLines

Boolean indicating whether to show grid lines in the table

table.tabColour

Color of the table tab in Excel. Hexadecimal color code (e.g., "#FF0000" for red) or a named color (e.g., "red"). Defaults to NULL, which means no tab colour.

table.fontName

Font name for the table text. Defaults to NULL, which means the default font will be used.

table.fontSize

Font size for the table text. Defaults to NULL, which means the default font size will be used.

table.fontColour

Font colour for the table text. Hexadecimal color code (e.g., "#FFFFFF" for white) or a named color (e.g., "white"). Defaults to NULL, which means no specific font colour will be applied.

table.bgFill

Background fill for the table cells. Hexadecimal color code. Defaults to NULL, which means no specific background fill will be applied.

table.fgFill

Foreground fill for the table cells. Hexadecimal color code. Defaults to NULL, which means no specific foreground fill will be applied.

table.halign

Horizontal alignment of the table text. Options are "left", "center", or "right". Defaults to NULL, which means the default alignment will be used.

table.valign

Vertical alignment of the table text. Options are "top", "middle", or "bottom". Defaults to NULL, which means the default alignment will be used.

table.wrapText

Boolean indicating whether to wrap text in the table cells. Defaults to FALSE, which means text will not be wrapped.

table.indent

Indentation for the table text. Numeric value indicating the number of spaces to indent. Defaults to NULL, which means no indentation will be applied.

table.locked

Boolean indicating whether the table cells should be locked for editing. Defaults to NULL, which means the default lock setting will be used.

table.hidden

Boolean indicating whether the table cells should be hidden. Defaults to NULL, which means the default visibility setting will be used.

table.decimalPrecision

Number of decimal places to display for numeric columns in the table. Defaults to NULL, which means no specific precision will be applied.

table.decimalCols

A character vector specifying which columns should have decimal precision applied. Defaults to NULL, which means no specific columns will be targeted.

table.lastRowBold

Boolean indicating whether the last row of the table should be bold. Defaults to NULL, which means the default setting will be used.

table.widths

A numeric vector specifying the widths of the columns in the table. Defaults to NULL, which means the default column widths will be used.

table.widthOffset

Numeric value indicating the offset to apply to the column widths. Defaults to NULL, which means no offset will be applied.

title.fontName

Font name for the table title. Defaults to NULL, which means the default font will be used.

title.fontSize

Font size for the table title. Defaults to NULL, which means the default font size will be used.

title.fontColour

Font colour for the table title. Hexadecimal color code. Defaults to NULL, which means no specific font colour will be applied.

title.border

Boolean indicating whether to apply a border to the table title. Defaults to NULL, which means no border will be applied.

title.borderColour

Border colour for the table title. Hexadecimal color code. Defaults to NULL, which means no specific border colour will be applied.

title.borderStyle

Border style for the table title. Options are "thin", "medium", "thick", etc. Defaults to NULL, which means no specific border style will be applied.

title.bgFill

Background fill for the table title. Hexadecimal color code. Defaults to NULL, which means no specific background fill will be applied.

title.fgFill

Foreground fill for the table title. Hexadecimal color code. Defaults to NULL, which means no specific foreground fill will be applied.

title.halign

Horizontal alignment of the table title text. Options are "left", "center", or "right". Defaults to NULL, which means the default alignment will be used.

title.valign

Vertical alignment of the table title text. Options are "top", "middle", or "bottom". Defaults to NULL, which means the default alignment will be used.

title.textDecoration

Text decoration for the table title. Options are "none", "bold", "italic", etc. Defaults to NULL, which means no specific text decoration will be applied.

title.wrapText

Boolean indicating whether to wrap text in the table title. Defaults to NULL, which means the default setting will be used.

title.textRotation

Text rotation for the table title. Numeric value indicating the angle of rotation in degrees. Defaults to NULL, which means no rotation will be applied.

title.indent

Indentation for the table title text. Numeric value indicating the number of spaces to indent. Defaults to NULL, which means no indentation will be applied.

title.heights

A numeric vector specifying the heights of the rows in the table title. Defaults to NULL, which means the default row heights will be used.

subtitle.fontName

Font name for the table subtitle. Defaults to NULL, which means the default font will be used.

subtitle.fontSize

Font size for the table subtitle. Defaults to NULL, which means the default font size will be used.

subtitle.fontColour

Font colour for the table subtitle. Hexadecimal color code. Defaults to NULL, which means no specific font colour will be applied.

subtitle.border

Boolean indicating whether to apply a border to the table subtitle. Defaults to NULL, which means no border will be applied.

subtitle.borderColour

Border colour for the table subtitle. Hexadecimal color code. Defaults to NULL, which means no specific border colour will be applied.

subtitle.borderStyle

Border style for the table subtitle. Options are "thin", "medium", "thick", etc. Defaults to NULL, which means no specific border style will be applied.

subtitle.bgFill

Background fill for the table subtitle. Hexadecimal color code. Defaults to NULL, which means no specific background fill will be applied.

subtitle.fgFill

Foreground fill for the table subtitle. Hexadecimal color code. Defaults to NULL, which means no specific foreground fill will be applied.

subtitle.halign

Horizontal alignment of the table subtitle text. Options are "left", "center", or "right". Defaults to NULL, which means the default alignment will be used.

subtitle.valign

Vertical alignment of the table subtitle text. Options are "top", "middle", or "bottom". Defaults to NULL, which means the default alignment will be used.

subtitle.textDecoration

Text decoration for the table subtitle. Options are "none", "bold", "italic", etc. Defaults to NULL, which means no specific text decoration will be applied.

subtitle.wrapText

Boolean indicating whether to wrap text in the table subtitle. Defaults to NULL, which means the default setting will be used.

subtitle.textRotation

Text rotation for the table subtitle. Numeric value indicating the angle of rotation in degrees. Defaults to NULL, which means no rotation will be applied.

subtitle.indent

Indentation for the table subtitle text. Numeric value indicating the number of spaces to indent. Defaults to NULL, which means no indentation will be applied.

subtitle.heights

A numeric vector specifying the heights of the rows in the table subtitle. Defaults to NULL, which means the default row heights will be used.

header.fontName

Font name for the table header. Defaults to NULL, which means the default font will be used.

header.fontSize

Font size for the table header. Defaults to NULL, which means the default font size will be used.

header.fontColour

Font colour for the table header. Hexadecimal color code. Defaults to NULL, which means no specific font colour will be applied.

header.border

Boolean indicating whether to apply a border to the table header. Defaults to NULL, which means no border will be applied.

header.borderColour

Border colour for the table header. Hexadecimal color code. Defaults to NULL, which means no specific border colour will be applied.

header.borderStyle

Border style for the table header. Options are "thin", "medium", "thick", etc. Defaults to NULL, which means no specific border style will be applied.

header.bgFill

Background fill for the table header. Hexadecimal color code. Defaults to NULL, which means no specific background fill will be applied.

header.fgFill

Foreground fill for the table header. Hexadecimal color code. Defaults to NULL, which means no specific foreground fill will be applied.

header.halign

Horizontal alignment of the table header text. Options are "left", "center", or "right". Defaults to NULL, which means the default alignment will be used.

header.valign

Vertical alignment of the table header text. Options are "top", "middle", or "bottom". Defaults to NULL, which means the default alignment will be used.

header.textDecoration

Text decoration for the table header. Options are "none", "bold", "italic", etc. Defaults to NULL, which means no specific text decoration will be applied.

header.wrapText

Boolean indicating whether to wrap text in the table header. Defaults to NULL, which means the default setting will be used.

header.textRotation

Text rotation for the table header. Numeric value indicating the angle of rotation in degrees. Defaults to NULL, which means no rotation will be applied.

header.indent

Indentation for the table header text. Numeric value indicating the number of spaces to indent. Defaults to NULL, which means no indentation will be applied.

header.heights

A numeric vector specifying the heights of the rows in the table header. Defaults to NULL, which means the default row heights will be used.

spanner.fontName

Font name for the table spanner (a spanning header). Defaults to NULL, which means the default font will be used.

spanner.fontSize

Font size for the table spanner. Defaults to NULL, which means the default font size will be used.

spanner.fontColour

Font colour for the table spanner. Hexadecimal color code. Defaults to NULL, which means no specific font colour will be applied.

spanner.border

Boolean indicating whether to apply a border to the table spanner. Defaults to NULL, which means no border will be applied.

spanner.borderColour

Border colour for the table spanner. Hexadecimal color code. Defaults to NULL, which means no specific border colour will be applied.

spanner.borderStyle

Border style for the table spanner. Options are "thin", "medium", "thick", etc. Defaults to NULL, which means no specific border style will be applied.

spanner.bgFill

Background fill for the table spanner. Hexadecimal color code. Defaults to NULL, which means no specific background fill will be applied.

spanner.fgFill

Foreground fill for the table spanner. Hexadecimal color code. Defaults to NULL, which means no specific foreground fill will be applied.

spanner.halign

Horizontal alignment of the table spanner text. Options are "left", "center", or "right". Defaults to NULL, which means the default alignment will be used.

spanner.valign

Vertical alignment of the table spanner text. Options are "top", "middle", or "bottom". Defaults to NULL, which means the default alignment will be used.

spanner.textDecoration

Text decoration for the table spanner. Options are "none", "bold", "italic", etc. Defaults to NULL, which means no specific text decoration will be applied.

spanner.wrapText

Boolean indicating whether to wrap text in the table spanner. Defaults to NULL, which means the default setting will be used.

spanner.textRotation

Text rotation for the table spanner. Numeric value indicating the angle of rotation in degrees. Defaults to NULL, which means no rotation will be applied.

spanner.indent

Indentation for the table spanner text. Numeric value indicating the number of spaces to indent. Defaults to NULL, which means no indentation will be applied.

spanner.heights

A numeric vector specifying the heights of the rows in the table spanner. Defaults to NULL, which means the default row heights will be used.

body.fontName

Font name for the table body. Defaults to NULL, which means the default font will be used.

body.fontSize

Font size for the table body. Defaults to NULL, which means the default font size will be used.

body.fontColour

Font colour for the table body. Hexadecimal color code. Defaults to NULL, which means no specific font colour will be applied.

body.numFmt

Number format for the table body. This can be a string representing a number format (e.g., "0.00" for two decimal places). Defaults to NULL, which means no specific number format will be applied.

body.border

Boolean indicating whether to apply a border to the table body. Defaults to NULL, which means no border will be applied.

body.borderColour

Border colour for the table body. Hexadecimal color code. Defaults to NULL, which means no specific border colour will be applied.

body.borderStyle

Border style for the table body. Options are "thin", "medium", "thick", etc. Defaults to NULL, which means no specific border style will be applied.

body.bgFill

Background fill for the table body. Hexadecimal color code. Defaults to NULL, which means no specific background fill will be applied.

body.fgFill

Foreground fill for the table body. Hexadecimal color code. Defaults to NULL, which means no specific foreground fill will be applied.

body.halign

Horizontal alignment of the table body text. Options are "left", "center", or "right". Defaults to NULL, which means the default alignment will be used.

body.valign

Vertical alignment of the table body text. Options are "top", "middle", or "bottom". Defaults to NULL, which means the default alignment will be used.

body.textDecoration

Text decoration for the table body. Options are "none", "bold", "italic", etc. Defaults to NULL, which means no specific text decoration will be applied.

body.wrapText

Boolean indicating whether to wrap text in the table body. Defaults to NULL, which means the default setting will be used.

body.textRotation

Text rotation for the table body. Numeric value indicating the angle of rotation in degrees. Defaults to NULL, which means no rotation will be applied.

body.indent

Indentation for the table body text. Numeric value indicating the number of spaces to indent. Defaults to NULL, which means no indentation will be applied.

body.heights

A numeric vector specifying the heights of the rows in the table body. Defaults to NULL, which means the default row heights will be used.

col_first.fontName

Font name for the first column of the table. Defaults to NULL, which means the default font will be used.

col_first.fontSize

Font size for the first column of the table. Defaults to NULL, which means the default font size will be used.

col_first.fontColour

Font colour for the first column of the table. Hexadecimal color code. Defaults to NULL, which means no specific font colour will be applied.

col_first.numFmt

Number format for the first column of the table. This can be a string representing a number format (e.g., "0.00" for two decimal places). Defaults to NULL, which means no specific number format will be applied.

col_first.border

Boolean indicating whether to apply a border to the first column of the table. Defaults to NULL, which means no border will be applied.

col_first.borderColour

Border colour for the first column of the table. Hexadecimal color code. Defaults to NULL, which means no specific border colour will be applied.

col_first.borderStyle

Border style for the first column of the table. Options are "thin", "medium", "thick", etc. Defaults to NULL, which means no specific border style will be applied.

col_first.bgFill

Background fill for the first column of the table. Hexadecimal color code. Defaults to NULL, which means no specific background fill will be applied.

col_first.fgFill

Foreground fill for the first column of the table. Hexadecimal color code. Defaults to NULL, which means no specific foreground fill will be applied.

col_first.halign

Horizontal alignment of the first column text. Options are "left", "center", or "right". Defaults to NULL, which means the default alignment will be used.

col_first.valign

Vertical alignment of the first column text. Options are "top", "middle", or "bottom". Defaults to NULL, which means the default alignment will be used.

col_first.textDecoration

Text decoration for the first column. Options are "none", "bold", "italic", etc. Defaults to NULL, which means no specific text decoration will be applied.

col_first.wrapText

Boolean indicating whether to wrap text in the first column. Defaults to NULL, which means the default setting will be used.

col_first.textRotation

Text rotation for the first column. Numeric value indicating the angle of rotation in degrees. Defaults to NULL, which means no rotation will be applied.

col_first.indent

Indentation for the first column text. Numeric value indicating the number of spaces to indent. Defaults to NULL, which means no indentation will be applied.

col_first.widths

A numeric vector specifying the widths of the first column. Defaults to NULL, which means the default column width will be used.

col_last.fontName

Font name for the last column of the table. Defaults to NULL, which means the default font will be used.

col_last.fontSize

Font size for the last column of the table. Defaults to NULL, which means the default font size will be used.

col_last.fontColour

Font colour for the last column of the table. Hexadecimal color code. Defaults to NULL, which means no specific font colour will be applied.

col_last.numFmt

Number format for the last column of the table. This can be a string representing a number format (e.g., "0.00" for two decimal places). Defaults to NULL, which means no specific number format will be applied.

col_last.border

Boolean indicating whether to apply a border to the last column of the table. Defaults to NULL, which means no border will be applied.

col_last.borderColour

Border colour for the last column of the table. Hexadecimal color code. Defaults to NULL, which means no specific border colour will be applied.

col_last.borderStyle

Border style for the last column of the table. Options are "thin", "medium", "thick", etc. Defaults to NULL, which means no specific border style will be applied.

col_last.bgFill

Background fill for the last column of the table. Hexadecimal color code. Defaults to NULL, which means no specific background fill will be applied.

col_last.fgFill

Foreground fill for the last column of the table. Hexadecimal color code. Defaults to NULL, which means no specific foreground fill will be applied.

col_last.halign

Horizontal alignment of the last column text. Options are "left", "center", or "right". Defaults to NULL, which means the default alignment will be used.

col_last.valign

Vertical alignment of the last column text. Options are "top", "middle", or "bottom". Defaults to NULL, which means the default alignment will be used.

col_last.textDecoration

Text decoration for the last column. Options are "none", "bold", "italic", etc. Defaults to NULL, which means no specific text decoration will be applied.

col_last.wrapText

Boolean indicating whether to wrap text in the last column. Defaults to NULL, which means the default setting will be used.

col_last.textRotation

Text rotation for the last column. Numeric value indicating the angle of rotation in degrees. Defaults to NULL, which means no rotation will be applied.

col_last.indent

Indentation for the last column text. Numeric value indicating the number of spaces to indent. Defaults to NULL, which means no indentation will be applied.

col_last.widths

A numeric vector specifying the widths of the last column. Defaults to NULL, which means the default column width will be used.

row_group.fontName

Font name for the row group header. Defaults to NULL, which means the default font will be used.

row_group.fontSize

Font size for the row group header. Defaults to NULL, which means the default font size will be used.

row_group.fontColour

Font colour for the row group header. Hexadecimal color code. Defaults to NULL, which means no specific font colour will be applied.

row_group.border

Boolean indicating whether to apply a border to the row group header. Defaults to NULL, which means no border will be applied.

row_group.borderColour

Border colour for the row group header. Hexadecimal color code. Defaults to NULL, which means no specific border colour will be applied.

row_group.borderStyle

Border style for the row group header. Options are "thin", "medium", "thick", etc. Defaults to NULL, which means no specific border style will be applied.

row_group.bgFill

Background fill for the row group header. Hexadecimal color code. Defaults to NULL, which means no specific background fill will be applied.

row_group.fgFill

Foreground fill for the row group header. Hexadecimal color code. Defaults to NULL, which means no specific foreground fill will be applied.

row_group.halign

Horizontal alignment of the row group header text. Options are "left", "center", or "right". Defaults to NULL, which means the default alignment will be used.

row_group.valign

Vertical alignment of the row group header text. Options are "top", "middle", or "bottom". Defaults to NULL, which means the default alignment will be used.

row_group.textDecoration

Text decoration for the row group header. Options are "none", "bold", "italic", etc. Defaults to NULL, which means no specific text decoration will be applied.

row_group.wrapText

Boolean indicating whether to wrap text in the row group header. Defaults to NULL, which means the default setting will be used.

row_group.textRotation

Text rotation for the row group header. Numeric value indicating the angle of rotation in degrees. Defaults to NULL, which means no rotation will be applied.

row_group.indent

Indentation for the row group header text. Numeric value indicating the number of spaces to indent. Defaults to NULL, which means no indentation will be applied.

row_group.widths

A numeric vector specifying the widths of the columns in the row group header. Defaults to NULL, which means the default column widths will be used.

source_note.fontName

Font name for the source note. Defaults to NULL, which means the default font will be used.

source_note.fontSize

Font size for the source note. Defaults to NULL, which means the default font size will be used.

source_note.fontColour

Font colour for the source note. Hexadecimal color code. Defaults to NULL, which means no specific font colour will be applied.

source_note.border

Boolean indicating whether to apply a border to the source note. Defaults to NULL, which means no border will be applied.

source_note.borderColour

Border colour for the source note. Hexadecimal color code. Defaults to NULL, which means no specific border colour will be applied.

source_note.borderStyle

Border style for the source note. Options are "thin", "medium", "thick", etc. Defaults to NULL, which means no specific border style will be applied.

source_note.bgFill

Background fill for the source note. Hexadecimal color code. Defaults to NULL, which means no specific background fill will be applied.

source_note.fgFill

Foreground fill for the source note. Hexadecimal color code. Defaults to NULL, which means no specific foreground fill will be applied.

source_note.halign

Horizontal alignment of the source note text. Options are "left", "center", or "right". Defaults to NULL, which means the default alignment will be used.

source_note.valign

Vertical alignment of the source note text. Options are "top", "middle", or "bottom". Defaults to NULL, which means the default alignment will be used.

source_note.textDecoration

Text decoration for the source note. Options are "none", "bold", "italic", etc. Defaults to NULL, which means no specific text decoration will be applied.

source_note.wrapText

Boolean indicating whether to wrap text in the source note. Defaults to NULL, which means the default setting will be used.

source_note.textRotation

Text rotation for the source note. Numeric value indicating the angle of rotation in degrees. Defaults to NULL, which means no rotation will be applied.

source_note.indent

Indentation for the source note text. Numeric value indicating the number of spaces to indent. Defaults to NULL, which means no indentation will be applied.

source_note.heights

A numeric vector specifying the heights of the rows in the source note. Defaults to NULL, which means the default row heights will be used.

footnotes.fontName

Font name for the footnotes. Defaults to NULL, which means the default font will be used.

footnotes.fontSize

Font size for the footnotes. Defaults to NULL, which means the default font size will be used.

footnotes.fontColour

Font colour for the footnotes. Hexadecimal color code. Defaults to NULL, which means no specific font colour will be applied.

footnotes.border

Boolean indicating whether to apply a border to the footnotes. Defaults to NULL, which means no border will be applied.

footnotes.borderColour

Border colour for the footnotes. Hexadecimal color code. Defaults to NULL, which means no specific border colour will be applied.

footnotes.borderStyle

Border style for the footnotes. Options are "thin", "medium", "thick", etc. Defaults to NULL, which means no specific border style will be applied.

footnotes.bgFill

Background fill for the footnotes. Hexadecimal color code. Defaults to NULL, which means no specific background fill will be applied.

footnotes.fgFill

Foreground fill for the footnotes. Hexadecimal color code. Defaults to NULL, which means no specific foreground fill will be applied.

footnotes.halign

Horizontal alignment of the footnotes text. Options are "left", "center", or "right". Defaults to NULL, which means the default alignment will be used.

footnotes.valign

Vertical alignment of the footnotes text. Options are "top", "middle", or "bottom". Defaults to NULL, which means the default alignment will be used.

footnotes.textDecoration

Text decoration for the footnotes. Options are "none", "bold", "italic", etc. Defaults to NULL, which means no specific text decoration will be applied.

footnotes.wrapText

Boolean indicating whether to wrap text in the footnotes. Defaults to NULL, which means the default setting will be used.

footnotes.textRotation

Text rotation for the footnotes. Numeric value indicating the angle of rotation in degrees. Defaults to NULL, which means no rotation will be applied.

footnotes.indent

Indentation for the footnotes text. Numeric value indicating the number of spaces to indent. Defaults to NULL, which means no indentation will be applied.

footnotes.heights

A numeric vector specifying the heights of the rows in the footnotes. Defaults to NULL, which means the default row heights will be used.

Value

A tsg object with the specified facade settings applied as attributes.

Examples

person_record |>
 generate_frequency(sex) |>
 add_facade(table.offsetRow = 2, table.offsetCol = 1)
#> # A tibble: 3 × 3
#>   category   frequency percent
#>   <int+lbl>      <int>   <dbl>
#> 1 1 [Male]        1516    52.0
#> 2 2 [Female]      1402    48.0
#> 3 0 [Total]       2918   100